Update to docbook-xslTNG 2.6.1
This updates to docbook xslTNG 2.6.1

Honestly I'm not entirely sure what mediaobject path
options are required, it's hard to understand.  This seems
to work.  Paths updated as required.

No need for .eps any more.

Prince hasn't updated for Debian 13, use a container

Reorganize the calling of the docbook wrapper script
to be a bit simpler.